fixed ssr issue

+++ src/app/shared/services/blog/blog.ts
@@ -1,7 +1,8 @@
-import { Injectable } from '@angular/core';
+import { Injectable, Inject, PLATFORM_ID } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { environment } from '../../../../environments/environment';
-import { Observable } from 'rxjs';
+import { Observable, of } from 'rxjs';
+import { isPlatformBrowser } from '@angular/common';
 
 @Injectable({
   providedIn: 'root',
@@ -9,7 +10,10 @@ import { Observable } from 'rxjs';
 export class BlogApiService {
   private apiUrl = environment.blogsApiBaseUrl;
 
-  constructor(private http: HttpClient) {}
+  constructor(
+    private http: HttpClient,
+    @Inject(PLATFORM_ID) private platformId: Object
+  ) {}
 
   getBlogs(): Observable<any> {
     return this.http.get(`${this.apiUrl}/get-blog-list`);
@@ -19,15 +23,25 @@ export class BlogApiService {
     return this.http.get(`${this.apiUrl}/get-blog/${id}`);
   }
 
+  // Only run in browser (to avoid SSR errors for mutations)
   updateBlog(id: string, blog: any): Observable<any> {
-    return this.http.put(`${this.apiUrl}/edit-blog/${id}`, blog);
+    if (isPlatformBrowser(this.platformId)) {
+      return this.http.put(`${this.apiUrl}/edit-blog/${id}`, blog);
+    }
+    return of(null); // SSR-safe fallback
   }
 
   createBlog(blog: any): Observable<any> {
-    return this.http.post(`${this.apiUrl}/create-blog`, blog);
+    if (isPlatformBrowser(this.platformId)) {
+      return this.http.post(`${this.apiUrl}/create-blog`, blog);
+    }
+    return of(null);
   }
 
   deleteBlog(id: string): Observable<any> {
-    return this.http.delete(`${this.apiUrl}/delete-blog/${id}`);
+    if (isPlatformBrowser(this.platformId)) {
+      return this.http.delete(`${this.apiUrl}/delete-blog/${id}`);
+    }
+    return of(null);
   }
 }
